Initial Commit
Fleshed out some of the primary classes:
* BreedingChamber - The house for all genetic algorithm functions.
* Chromosome - The primary storage unit for chromosome values.
* ChromosomeFactory - A factory interface for generating chromosomes.
* ConvectionStorage - Used to store chromosomes in such a way that
allocated memory can be reused throughout the entire GA process.
* FitnessBlock - This is a container class for the fitness values.

Created the outlines for the other necessary classes:
* FileObject
* SpiceFileGenerator
* SpiceFileTemplate

Implemented two significant optimizations:
1. The FitnessBlock class was added to allow statistics on the fitness
data to be taken as the it is added to the data block. For now, the only
analysis is the sum of the fitness values. I took this a step further my
implementing a summation array that stores the sum of the fitness values
up to each index in the array.
------------------------------------------------------
Formal Definition:
------------------------------------------------------
||SA = summation array
||FIT = array storing the fitness values
||SA.length == FIT.length
||For each index i in SA, SA[i] = sum[k=0,i](FIT[k])
------------------------------------------------------
Hence, the final value in the summation array is the total sum of the
fitness values. This array is naturally already sorted upon creation.
That allows a modified binary search to be implemented to improve the
speed of the GA roulette selection. Because the fitness values will be
added by a number of threads, I must wait until all values are added
before filling the summation array.

2. The ConvectionStorage class was added to allow the easy reuse of
pre-allocated memory in the use of Chromosome storage. Two arrays are
rotated. One is the working set, the set of old chromosomes to be
reused, and the other is the historical set, the set of newly fitted
Chromosomes.

#No testing has been done up to this point, but will be completed in
future commits.