Randomized Operations and Simulated Annealing
* Merged FitnessBlock logic into ConvectionStorage.
* Chromosomes now store their personal fitness value.
* Created RegressiveConvectionStorage to allow checked out chromosomes
to be returned to the queue.
* Created GeneticOperation interface.
* Created SAGeneticOperation abstract class to handle simulated
annealing operations.
* Created SAController to control the temperature between all
SAGeneticOperation's.
* Created FittingRoom as the primary utility for measuring the fitness
of chromosomes.
* Created ChromosomeQueue to allow operation access to the
RegressiveConvectionStorage queue.
* Created a Utils class to store commonly useful functions.
* Created OperationGroup to store a set of operations and related
data/functions.
* GATest is now obsolete.
* FitnessBlock is not obsolete.
* The GeneticOperation's operate method now has updated parameters.
* Rearranged packages.

Special Notes:
* GeneticOperation.operate now has a Boolean parameter that specifies if
chromosomes should be fitted in the operation. Simulated annealing
operations often need to fit some or all of the children anyway, but
this may be helpful for non-SA operations when multiple operations are
being chained. This makes it possible to avoid unnecessary fitness
calculations.
* Operations should finalize chromosomes as an initial step, this
minimizes the potential amount of waiting time threads may experience
toward the end of population generations.
* ConvectionStorage.getChromosome has now been made thread safe. The
regressive subclass will suspend until either the numRemaining decreases
to 0, at which time it returns null, or a chromosome becomes available.
* I have determined that the collection of working chromosomes must be
done atomically as it is the most efficient way to deal with race
conditions.
* I have added quite a bit of code for training neural networks.

Future Considerations:
* Currently, if there are not enough chromosomes remaining compared to
the required number of output chromosomes of a selected operation, a new
operation is chosen. It can potentially be the case that the operation's
final number of chromosomes is within the bounds of the remaining
chromosome but still is not selected due to this constraint. This can be
resolved by having some sort of reserve system. This would mean
rebuilding the working set entirely and leaving an adequate number of
extra chromosomes within the queue.